Upgrade Protractor dependency to include v4
This replaces grunt-protractor-runner with a shell command so any
Protractor version later than the currently-supported v2.2.0 can be
used. (See also https://github.com/teerapap/grunt-protractor-runner/pull/164)

Also I don't think Protractor needs to be installed as
a dependency, so this moves it to a peer-dependency whilst also
including it as a dev-dependency so it can be run in the tests.