Use console.log instead of throw when testability is missing
Currently if window.testability is missing it shows this at the end of
the test: `Failure during waitForPromise: undefined`.  From what
I've read there's not a great way to return an error from
`browser.executeAsyncScript` and so I've changed it to just
`console.log` which shows up nicely in the protractor output.